Issue #11: improve types in file i/o functions
Use open(2)/read(2)/write(2) instead of fopen(3)/fread(3)/fwrite(3) in
several places to avoid unnecessary buffering

Fix a bug in Rhizome HTTP add's handling of unconfigured manifest
template file

Improve some debug and error logging for file i/o

+++ net.c
@@ -77,11 +77,11 @@ ssize_t _write_all(int fd, const void *buf, size_t len, struct __sourceloc __whe
 {
   ssize_t written = write(fd, buf, len);
   if (written == -1)
-    return WHYF_perror("write_all: write(%d,%p %s,%lu)",
-	fd, buf, alloca_toprint(30, buf, len), (unsigned long)len);
+    return WHYF_perror("write_all: write(%d,%p %s,%zu)",
+	fd, buf, alloca_toprint(30, buf, len), len);
   if (written != len)
-    return WHYF_perror("write_all: write(%d,%p %s,%lu) returned %ld",
-	fd, buf, alloca_toprint(30, buf, len), (unsigned long)len, (long)written);
+    return WHYF_perror("write_all: write(%d,%p %s,%zu) returned %zd",
+	fd, buf, alloca_toprint(30, buf, len), len, (size_t)written);
   return written;
 }
